chore: add skip entry for renamed e2e test
This test suite file was moved in https://github.com/vercel/next.js/pull/77630.

This results in it showing up as failed in new reports:
https://682e414a03ba12f55c5f417b--runtime-e2e-report.netlify.app/.

Example run here:
https://github.com/opennextjs/opennextjs-netlify/actions/runs/15163298543#summary-42664185436.

I also clarified the `reason` field.

We could do something gross to expect the Netlify default 404 page or configure a custom 404 page
that mimics Vercel's... but I don't think it's worth it.